app/api/v2/requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.0.4
autoenv==1.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.5.0
cryptography==2.3.1
docopt==0.6.2
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
python-dotenv==0.9.1
pytz==2018.5
requests==2.19.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11
